requirements.txt
boto3>=1.34.0
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.5.0
PyYAML>=6.0.1
streamlit>=1.36.0
plotly>=5.22.0
tiktoken>=0.7.0
requests>=2.32.0
tenacity>=8.3.0
python-dotenv>=1.0.1
sqlite-utils>=3.36
tqdm>=4.66.0
statistics; python_version < '3.8'

